[refactor] 기능별 패키지 구조 변경 및 일관성 있는 클래스 이름 변경
