+++ backend/src/main/java/com/ll/hotel/domain/member/member/service/MemberService.java
@@ -8,7 +8,7 @@ import com.ll.hotel.domain.member.member.repository.MemberRepository;
 import com.ll.hotel.domain.member.member.type.MemberStatus;
 import com.ll.hotel.global.exceptions.ServiceException;
 import com.ll.hotel.global.rsData.RsData;
-import com.ll.hotel.global.security.oauth2.CustomOAuth2JwtProperties;
+import com.ll.hotel.global.jwt.dto.JwtProperties;
 import com.ll.hotel.global.security.oauth2.entity.OAuth;
 import com.ll.hotel.global.security.oauth2.repository.OAuthRepository;
 import jakarta.validation.Valid;
@@ -32,7 +32,7 @@ public class MemberService {
     private final AuthTokenService authTokenService;
     private final RefreshTokenService refreshTokenService;
     private final RedisTemplate<String, String> redisTemplate;
-    private final CustomOAuth2JwtProperties jwtProperties;
+    private final JwtProperties jwtProperties;
     private static final Logger log = LoggerFactory.getLogger(MemberService.class);
     private static final String LOGOUT_PREFIX = "LOGOUT:";
