+++ backend/src/main/java/com/ll/hotel/global/jwt/dto/GeneratedToken.java
@@ -1,4 +1,4 @@
-package com.ll.hotel.global.security.dto;
+package com.ll.hotel.global.jwt.dto;
 
 import lombok.Builder;
