+++ backend/src/main/java/com/ll/hotel/global/jwt/dto/JwtProperties.java
@@ -1,4 +1,4 @@
-package com.ll.hotel.global.security.oauth2;
+package com.ll.hotel.global.jwt.dto;
 
 import lombok.Data;
 import org.springframework.boot.context.properties.ConfigurationProperties;
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Component;
 @Data
 @Component
 @ConfigurationProperties(prefix = "custom.jwt")
-public class CustomOAuth2JwtProperties {
+public class JwtProperties {
     private String secret;
     private long accessTokenExpiration;
     private long refreshTokenExpiration;
