+++ backend/src/main/java/com/ll/hotel/global/security/oauth2/CustomOAuth2AuthenticationSuccessHandler.java
@@ -3,7 +3,7 @@ package com.ll.hotel.global.security.oauth2;
 
 import com.ll.hotel.domain.member.member.service.AuthTokenService;
 import com.ll.hotel.domain.member.member.service.MemberService;
-import com.ll.hotel.global.security.dto.SecurityUser;
+import com.ll.hotel.global.security.oauth2.dto.SecurityUser;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
